Section 6.75

+++ src/Resolvers/Mutation.js
@@ -1,6 +1,8 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
+//password: red12345
 const Mutation = {
     //1. Take in Password 
     //2. Validate Password
@@ -27,41 +29,71 @@ const Mutation = {
         }
 
     }, 
-
-    async deleteUser(parent, args, { prisma }, info) {
-        const userExists = await prisma.exists.User({ id: args.id})
-
-        if(!userExists) {
-            throw new Error('User not found')
+    async login(parent, args, { prisma }, info) {
+        const user = await prisma.query.user({ where: {
+            email: args.data.email }
+        })
+        
+        if(!user) {
+            throw new Error('Unable to log in')
         }
 
-        return prisma.mutation.deleteUser({ where: { id : args.id }}, info)
+        const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+        if(!isMatch) {
+            throw new Error('Unable to log in')
+        }
+
+        return {
+            user,
+            token: jwt.sign({userId: user.id}, 'thisisasecret')
+        }
+    },
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        return prisma.mutation.deleteUser({ 
+            where: { id : userId }}, info)
         
     },
 
-    async updateUser(parent, args , { prisma }, info) {
-
+    async updateUser(parent, args , { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             }, 
             data: args.data
         }, info)
     }, 
-    async createPost(parent, args, { prisma }, info) {
+    async createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.createPost({data: {
             title: args.data.title,
             body: args.data.body,
             published: args.data.published,
             author: {
                 connect: {
-                    id: args.data.author
+                    id: userId
                 }
             }
         }}, info)
 
     }, 
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
@@ -69,7 +101,20 @@ const Mutation = {
         }, info)
     },
 
-    updatePost(parent, args, { prisma}, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to update post')
+        }
+
         return prisma.mutation.updatePost({
             data: {
                 title: args.data.title,
@@ -84,13 +129,16 @@ const Mutation = {
     },
 
     
-    createComment(parent, args, { prisma }, info) {
+    createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        
+        
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -102,14 +150,40 @@ const Mutation = {
         }, info)
 
     }, 
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id, 
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
